+++ crates/sevenmark_parser/src/parser/macro/mod.rs
@@ -1,4 +1,5 @@
 pub mod macro_age;
+pub mod macro_clear;
 pub mod macro_footnote;
 pub mod macro_newline;
 pub mod macro_now;
@@ -6,6 +7,7 @@ pub mod macro_null;
 pub mod macro_variable;
 
 pub use macro_age::*;
+pub use macro_clear::*;
 pub use macro_footnote::*;
 pub use macro_newline::*;
 pub use macro_now::*;